+++ migrations/sql/02-methods.sql
@@ -6,8 +6,8 @@
 -- ****************************************************
 
 CREATE TABLE methods (
-    id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    problem_id INTEGER REFERENCES problems(id),
+    id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    problem_id BIGINT REFERENCES problems(id),
     name VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
     example TEXT NOT NULL
